+++ src/main/java/com/swyp/index/application/dto/EmailRequest.java
@@ -1,0 +1,10 @@
+package com.swyp.index.application.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record EmailRequest(
+        @NotBlank @Email
+        String email
+) {
+}
