+++ src/main/java/com/swyp/index/application/dto/EmailVerificationRequest.java
@@ -1,0 +1,13 @@
+package com.swyp.index.application.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record EmailVerificationRequest(
+        @NotBlank @Email
+        String email,
+
+        @NotBlank
+        String code
+) {
+}
