+++ src/main/java/com/swyp/index/application/dto/ErrorResponse.java
@@ -1,0 +1,4 @@
+package com.swyp.index.application.dto;
+
+public record ErrorResponse(String message) {
+}
