+++ src/main/java/com/swyp/index/application/dto/LoginRequest.java
@@ -1,0 +1,12 @@
+package com.swyp.index.application.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginRequest(
+        @NotBlank @Email
+        String email,
+        @NotBlank
+        String password
+) {
+}
