--- src/main/java/com/swyp/index/config/JwtUtil.java
@@ -1,101 +1,0 @@
-package com.swyp.index.config;
-
-import io.jsonwebtoken.security.Keys;
-import io.jsonwebtoken.*;
-import jakarta.annotation.PostConstruct;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import org.springframework.util.StringUtils;
-
-import javax.crypto.SecretKey;
-import java.security.Key;
-import java.util.Base64;
-import java.util.Date;
-
-@Slf4j
-@Component
-public class JwtUtil {
-    public static final String AUTHORIZATION_HEADER = "Authorization";
-    public static final String BEARER_PREFIX = "Bearer ";
-
-    @Value("${jwt.secret}")
-    private String secretKey;
-
-    @Value("${jwt.access-token-expiration-ms}")
-    private long accessTokenExpirationMinutes;
-
-    @Value("${jwt.refresh-token-expiration-ms}")
-    private long refreshTokenExpirationDays;
-
-    private SecretKey key;
-
-    @PostConstruct
-    public void init() {
-        byte[] bytes = Base64.getDecoder().decode(secretKey);
-        key = Keys.hmacShaKeyFor(bytes);
-    }
-
-    // 액세스 토큰 생성
-    public String generateAccessToken(String email) {
-        Date now = new Date();
-        Date expiration = new Date(now.getTime() + accessTokenExpirationMinutes * 60 * 1000);
-
-        return Jwts.builder()
-                .setSubject(email)
-                .setIssuedAt(now)
-                .setExpiration(expiration)
-                .signWith(key)
-                .compact();
-    }
-
-    // 리프레시 토큰 생성
-    public String generateRefreshToken(String email) {
-        Date now = new Date();
-        Date expiration = new Date(now.getTime() + refreshTokenExpirationDays * 24 * 60 * 60 * 1000);
-
-        return Jwts.builder()
-                .setSubject(email)
-                .setIssuedAt(now)
-                .setExpiration(expiration)
-                .signWith(key)
-                .compact();
-    }
-
-    // HTTP 요청에서 토큰 추출
-    public String resolveToken(HttpServletRequest request) {
-        String bearerToken = request.getHeader(AUTHORIZATION_HEADER);
-        if (StringUtils.hasText(bearerToken) && bearerToken.startsWith(BEARER_PREFIX)) {
-            return bearerToken.substring(7);
-        }
-        return null;
-    }
-
-    // 토큰 유효성 검증
-    public boolean validateToken(String token) {
-        try {
-            Jwts.parser().verifyWith(key).build().parseSignedClaims(token);
-            return true;
-        } catch (SecurityException | MalformedJwtException e) {
-            log.error("Invalid JWT signature, 유효하지 않는 JWT 서명 입니다.");
-        } catch (ExpiredJwtException e) {
-            log.error("Expired JWT token, 만료된 JWT token 입니다.");
-        } catch (UnsupportedJwtException e) {
-            log.error("Unsupported JWT token, 지원되지 않는 JWT 토큰 입니다.");
-        } catch (IllegalArgumentException e) {
-            log.error("JWT claims is empty, 잘못된 JWT 토큰 입니다.");
-        }
-        return false;
-    }
-
-    // 토큰에서 사용자 정보(이메일) 추출
-    public Claims getUserInfoFromToken(String token) {
-        return Jwts.parser().verifyWith(key).build().parseSignedClaims(token).getPayload();
-    }
-
-    public String getEmailFromToken(String token) {
-        return getUserInfoFromToken(token).getSubject();
-    }
-
-}
