--- src/main/java/com/swyp/index/config/WebClientConfig.java
@@ -1,13 +1,0 @@
-package com.swyp.index.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-    @Bean
-    public WebClient webClient(){
-        return WebClient.builder().build();
-    }
-}
