+++ src/main/java/com/swyp/index/domain/ProviderType.java
@@ -1,0 +1,5 @@
+package com.swyp.index.domain;
+
+public enum ProviderType {
+    LOCAL, KAKAO, GOOGLE, NAVER
+}
