--- src/main/java/com/swyp/index/dto/EmailRequest.java
@@ -1,10 +1,0 @@
-package com.swyp.index.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record EmailRequest(
-        @NotBlank @Email
-        String email
-) {
-}
