--- src/main/java/com/swyp/index/dto/ErrorResponse.java
@@ -1,4 +1,0 @@
-package com.swyp.index.dto;
-
-public record ErrorResponse(String message) {
-}
