--- src/main/java/com/swyp/index/dto/LoginRequest.java
@@ -1,12 +1,0 @@
-package com.swyp.index.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginRequest(
-        @NotBlank @Email
-        String email,
-        @NotBlank
-        String password
-) {
-}
