--- src/main/java/com/swyp/index/entity/ProviderType.java
@@ -1,5 +1,0 @@
-package com.swyp.index.entity;
-
-public enum ProviderType {
-    LOCAL, KAKAO, GOOGLE, NAVER
-}
