--- src/main/java/com/swyp/index/entity/Role.java
@@ -1,14 +1,0 @@
-package com.swyp.index.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-    USER("ROLE_USER", "일반 사용자"),
-    ADMIN("ROLE_ADMIN", "관리자");
-
-    private final String key;
-    private final String title;
-}
