--- src/main/java/com/swyp/index/exception/DuplicateEmailException.java
@@ -1,7 +1,0 @@
-package com.swyp.index.exception;
-
-public class DuplicateEmailException extends RuntimeException{
-    public DuplicateEmailException(String message){
-        super(message);
-    }
-}
