--- src/main/java/com/swyp/index/exception/DuplicateNicknameException.java
@@ -1,7 +1,0 @@
-package com.swyp.index.exception;
-
-public class DuplicateNicknameException extends RuntimeException{
-    public DuplicateNicknameException(String message){
-        super(message);
-    }
-}
