--- src/main/java/com/swyp/index/exception/GlobalExceptionHandler.java
@@ -1,51 +1,0 @@
-package com.swyp.index.exception;
-
-import com.swyp.index.dto.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import javax.security.auth.login.LoginException;
-
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-
-    //회원가입 시 이메일 중복 예외 처리
-    @ExceptionHandler(DuplicateEmailException.class)
-    public ResponseEntity<ErrorResponse> handleDuplicateEmailException(DuplicateEmailException ex){
-        ErrorResponse response = new ErrorResponse(ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.CONFLICT);
-    }
-
-    //회원가입 시 닉네임 중복 예외 처리
-    @ExceptionHandler(DuplicateNicknameException.class)
-    public ResponseEntity<ErrorResponse> handleDuplicateNicknameException(DuplicateNicknameException ex){
-        ErrorResponse response = new ErrorResponse(ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.CONFLICT);
-    }
-
-    //JWT 관련 예외 처리
-    @ExceptionHandler(TokenException.class)
-    public ResponseEntity<ErrorResponse> handleTokenException(TokenException ex){
-        ErrorResponse response = new ErrorResponse(ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.UNAUTHORIZED);
-    }
-
-    //DTO의 @Valid 유효성 검사에 실패했을 때 발생하는 예외 처리
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<ErrorResponse> handleValidationExceptions(MethodArgumentNotValidException ex){
-        String errorMessage = ex.getBindingResult().getAllErrors().get(0).getDefaultMessage();
-        ErrorResponse response = new ErrorResponse(errorMessage);
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-
-    //인증 코드 불일치 등 인증 자격 증명 관련 예외 처리
-    @ExceptionHandler(BadCredentialsException.class)
-    public ResponseEntity<ErrorResponse> handleBadCredentialsException(BadCredentialsException ex){
-        ErrorResponse response = new ErrorResponse(ex.getMessage());
-        return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-    }
-}
