--- src/main/java/com/swyp/index/exception/TokenException.java
@@ -1,7 +1,0 @@
-package com.swyp.index.exception;
-
-public class TokenException extends RuntimeException{
-    public TokenException(String message){
-        super(message);
-    }
-}
