+++ src/main/java/com/swyp/index/global/exception/DuplicateEmailException.java
@@ -1,0 +1,7 @@
+package com.swyp.index.global.exception;
+
+public class DuplicateEmailException extends RuntimeException{
+    public DuplicateEmailException(String message){
+        super(message);
+    }
+}
