+++ src/main/java/com/swyp/index/global/exception/DuplicateNicknameException.java
@@ -1,0 +1,7 @@
+package com.swyp.index.global.exception;
+
+public class DuplicateNicknameException extends RuntimeException{
+    public DuplicateNicknameException(String message){
+        super(message);
+    }
+}
