+++ src/main/java/com/swyp/index/global/exception/TokenException.java
@@ -1,0 +1,7 @@
+package com.swyp.index.global.exception;
+
+public class TokenException extends RuntimeException{
+    public TokenException(String message){
+        super(message);
+    }
+}
